+++ tests/test_live.py
@@ -8,7 +8,13 @@ from email.message import EmailMessage
 
 import pytest
 
-from aiosmtplib import SMTP, SMTPAuthenticationError, SMTPStatus, send
+from aiosmtplib import (
+    SMTP,
+    SMTPAuthenticationError,
+    SMTPSenderRefused,
+    SMTPStatus,
+    send,
+)
 
 
 pytestmark = [
@@ -35,7 +41,6 @@ async def test_starttls_gmail():
         await client.login("test", "test")
 
 
-@pytest.mark.asyncio()
 async def test_qq_login():
     client = SMTP(hostname="smtp.qq.com", port=587, use_tls=False)
     await client.connect(timeout=2.0)
@@ -62,3 +67,18 @@ async def test_office365_auth_send():
             password="test",
             username="test",
         )
+
+
+async def test_office365_skip_login():
+    message = EmailMessage()
+    message["From"] = "user@mydomain.com"
+    message["To"] = "somebody@example.com"
+    message["Subject"] = "Hello World!"
+    message.set_content("Sent via aiosmtplib")
+
+    smtp_client = SMTP("smtp.office365.com", 587)
+    await smtp_client.connect()
+    await smtp_client.starttls()
+    # skip login, which is required
+    with pytest.raises(SMTPSenderRefused):
+        await smtp_client.send_message(message)