+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 1.1.5 (unreleased)
 ------------------
 
+- Bugfix: avoid raising asyncio.CancelledError on connection lost
+
 
 1.1.4
 -----